move thread.rs into concurrency
