+++ src/concurrency/mod.rs
@@ -1,5 +1,6 @@
 pub mod data_race;
 mod range_object_map;
 pub mod sync;
+pub mod thread;
 mod vector_clock;
 pub mod weak_memory;
